Modified trajectory visualizer

+++ trajectory_visualizer.py
@@ -10,6 +10,8 @@ import os
 import torch
 from argparse import Namespace
 from car_utils import get_car_path, get_radius, get_car_line_mark, WHEEL_STEER_RATIO
+from tqdm import tqdm
+
 
 FILTER_NEGATIVE_POINTS = True
 FILTER_NONMONOTONIC_POINTS = True
@@ -324,7 +326,7 @@ def normalize(tensor):
 
 
 def compute_score(overlay, predicted):
-    return predicted[overlay != 0].mean()
+    return predicted[overlay != 0].sum()
 
 
 def compute_miou(overlay, predicted):
@@ -348,7 +350,7 @@ def main_revised():
     cfg.__dict__ = cfg_i
 
     tv = TrajectoryVisualizer(cfg)
-    num = 1001
+    num = 501
     max_wheel_angle = np.rad2deg(np.arctan(CAR_L / MIN_TURNING_RADIUS))
     angles = np.linspace(-MAX_STEER, MAX_STEER, num)
     r = 1
@@ -378,19 +380,28 @@ def main_revised():
     soft_labels_dir = '/home/nemodrive/workspace/andreim/awesome-semantic-segmentation-pytorch/runs/pred_tensor_all/deeplabv3_resnet50_upb/'
     hard_labels_dir = '/home/nemodrive/workspace/andreim/awesome-semantic-segmentation-pytorch/runs/pred_pic_hard_all/deeplabv3_resnet50_upb/'
 
-    save_dir_soft = '/home/nemodrive/workspace/andreim/self_supervised_steering_results/soft'
-    save_dir_hard = '/home/nemodrive/workspace/andreim/self_supervised_steering_results/hard'
+    save_dir_soft = '/home/nemodrive/workspace/andreim/self_supervised_steering_results/soft_{}'.format(num)
+    save_dir_hard = '/home/nemodrive/workspace/andreim/self_supervised_steering_results/hard_{}'.format(num)
 
     test_dirs = os.listdir(test_dir)
 
-    for d in sorted(test_dirs):
+    for d in tqdm(sorted(test_dirs)):
         if '.txt' not in d:
             dir_path = os.path.join(test_dir, d)
             test_files = os.listdir(dir_path)
             cam_file = os.path.join(dir_path, 'cam.txt')
             camera_matrix = np.loadtxt(cam_file)
-            for f in sorted(test_files):
-                print(f)
+
+            save_path_soft = os.path.join(save_dir_soft, d)
+            save_path_hard = os.path.join(save_dir_hard, d)
+
+            if not os.path.exists(save_path_soft):
+                os.makedirs(save_path_soft)
+
+            if not os.path.exists(save_path_hard):
+                os.makedirs(save_path_hard)
+
+            for f in tqdm(sorted(test_files)):
                 if '.txt' not in f:
                     file_path = os.path.join(dir_path, f)
 
@@ -399,33 +410,34 @@ def main_revised():
                     sequence_frame_path_soft = os.path.join(soft_labels_dir, sequence_frame).replace('.png', '.pt')
                     sequence_frame_path_hard = os.path.join(hard_labels_dir, sequence_frame)
 
-                    sequence_frame_soft = normalize(torch.load(sequence_frame_path_soft)).cpu().numpy()
-                    sequence_frame_hard = cv2.imread(sequence_frame_path_hard)[:, :, 1]
+                    if os.path.exists(sequence_frame_path_soft) and os.path.exists(sequence_frame_path_hard):
+                        sequence_frame_soft = normalize(torch.load(sequence_frame_path_soft)).cpu().numpy()
+                        sequence_frame_hard = cv2.imread(sequence_frame_path_hard)[:, :, 1]
 
-                    frame_results_soft = []
-                    frame_results_hard = []
+                        frame_results_soft = []
+                        frame_results_hard = []
 
-                    for angle in angles:
-                        image, overlay = update_steer(file_path, camera_matrix, angle)
-                        image = cv2.addWeighted(image, 1, np.repeat(sequence_frame_hard[:, :, None], 3, axis=2), 1, 0)
+                        for angle in angles:
+                            image, overlay = update_steer(file_path, camera_matrix, angle)
+                            image = cv2.addWeighted(image, 1, np.repeat(sequence_frame_hard[:, :, None], 3, axis=2), 1, 0)
 
-                        # cv2.imshow('image', image)
-                        # cv2.waitKey(0)
+                            # cv2.imshow('image', image)
+                            # cv2.waitKey(0)
 
-                        overlay = overlay[:, :, 1]
-                        overlay = overlay / overlay.max()
+                            overlay = overlay[:, :, 1]
+                            overlay = overlay / overlay.max()
 
-                        soft_score = compute_score(overlay, sequence_frame_soft)
-                        miou = compute_miou(overlay, sequence_frame_hard)
+                            soft_score = compute_score(overlay, sequence_frame_soft)
+                            miou = compute_miou(overlay, sequence_frame_hard)
 
-                        frame_results_soft.append(soft_score)
-                        frame_results_hard.append(miou)
+                            frame_results_soft.append(soft_score)
+                            frame_results_hard.append(miou)
 
-                    res_soft = np.array([angles, frame_results_soft])
-                    res_hard = np.array([angles, frame_results_hard])
+                        res_soft = np.array([angles, frame_results_soft])
+                        res_hard = np.array([angles, frame_results_hard])
 
-                    np.save(os.path.join(save_dir_soft, sequence_frame).replace('.png', '.npy'), res_soft)
-                    np.save(os.path.join(save_dir_hard, sequence_frame).replace('.png', '.npy'), res_hard)
+                        np.save(os.path.join(save_dir_soft, sequence_frame).replace('.png', '.npy'), res_soft)
+                        np.save(os.path.join(save_dir_hard, sequence_frame).replace('.png', '.npy'), res_hard)
 
 if __name__ == "__main__":
     main_revised()
